mering function with unit tests
